mclient: avoid "ghost" results in batch mode
When processing files without interactive mode (-i or as implied by -e)
"ghost" results could be printed at certain points in the query output.

The "ghost" result was in fact the last result from a block of data read
by mclient, that was repeated without any data results for the next
block of data read.  This is caused by calling mapi_needmore followed by
mapi_next_result, which is essential to move the current result pointer
on to the next in line.  Since mapi_next_result returns what
mapi_needmore returns, but also advances the result pointer,
mapi_needmore is simply not necessary, and due to the logic, causes the
previous result to be printed again in the format_result loop.