Fix alignment issue.
On 64 bit Windows mserver5 wouldn't start in the default branch.  The
reason for that was that setjmp needs a 16-byte-aligned pointer for
the setjmp buffer.  Malloc knows that and always returns
16-byte-aligned pointers, but GDKmalloc added 8 to that, so the
pointer setjmp got was only 8-byte-aligned.

We now add two times sizeof(void) to the pointer that malloc returns
which will keep the alignment on a 16 byte boundary on 64 bit
Windows).

Note that on 32 bit Windows malloc returns 8 byte aligned pointers,
and on 64 bit Linux malloc returns 16 byte aligned pointers.  This
means this change will keep those alignments in tact on those
platforms.

Since this could potentially also be an issue in the Aug2011 branch,
the bug is fixed here.