DefBlk: avoid using strncat for Darwin9 (crashes)
Use memcpy instead of strncat.  In this case should be cheaper as well,
since we know the size already.  The limit for strncat was off by one,
since strncat adds the trailing \0 always, so could theoretically be
at size + 2.  This, however was not the problem why strncat internally
fails on its checks on Darwin 9.

With this change, a debug-compiled Mx can unpack gdk sources on Darwin 9.