modpath: don't use dirname twice on the same pointer
At least on Darwin, dirname apparently does a copy to a static location,
and hence valgrind complains that the strpy has equal source and
destination pointers.  Avoid doing this (even though it shouldn't harm
in this case), and simply use strrchr to eat away the two last
components of the path.