JDBC: set SQLSTATE for SQLExceptions thrown
Since the server doesn't provide us with any SQLSTATE information, the
usefulness is limited to exceptions generated by the JDBC driver itself.
The set of vendor-specific exceptions and their descriptions that are
used in the implementation is listed in the SQLSTATEs file.  For all
server originated exceptions the generic data exception state (22000) is
reported.