opt_mitosis: more conservative estimation of piece size
Instead of assuming an operator with 2 arguments and 1 result,
that each have a (0-byte) void head and an 8-byte lng tail,
we now assume that both arguments and the result also have
an 8-byte lng (or oid for that matter) head.

(
As compromise, we could alternatively assume only the result
to have an 8-byte lng/oid head, i.e., replace the "6" by a "4".
)