json: do not force array or object at top level
Prepare the way for jaql functions, which will get their arguments as
JSON structs, and return a JSON struct as well.  Since, e.g. count()
will return a number, but range() an array, allow simple values to be
represented in a JSON struct, since there is no technical limitation not
to.