added MAL optimizer template as optimizer "show-case"
This optimizer is mainly for demonstrating how to implement a MAL
optimizer in priciple.

Purely for that purpose, the optimizer's task is to recognize MAL code
patterns like

 ...
 v3 := sql.append( ..., ..., ..., ..., v0 );
 ...
 v5 := ... v0 ...;
 ...

i.e., an sql.append() statement that is eventually followed by some other
statement later on in the MAL program that uses the same v0 BAT as
argument as the sql.append() statement does,

and transform them into

 ...
 v1 := aggr.count( v0 );
 v2 := algebra.slice( v0, 0, v1 );
 v3 := sql.append( ..., ..., ..., ..., v2 );
 ...
 v5 := ... v0 ...;
 ...

i.e., handing a BAT view v2 of BAT v0 as argument to the sql.append()
statement, rather than the original BAT v0.

As a refinement, patterns like

 ...
 v3 := sql.append( ..., ..., ..., ..., v0 );
 v4 := aggr.count( v0 );
 ...
 v5 := ... v0 ...;
 ...

are transformed into

 ...
 v4 := aggr.count( v0 );
 v2 := algebra.slice( v0, 0, v4 );
 v3 := sql.append( ..., ..., ..., ..., v2 );
 ...
 v5 := ... v0 ...;
 ...