unwrap: give explicit type, can be deduced from unwraptype
Because MAL does type analysis and resolution before execution, we need
to allow the type system to deduce the type of the returned BAT by
unwrap.  For that, we need an argument that's the return can be deduced
from.  That argument can only be given if one knows what unwrap would
normally return without argument.  Hence unwraptype() to return that
type, as a string.