dumpvariabletransformation: fix order problem for var/var calculations
transform00 revealed an odd calculation error when int values had to be
upcasted to dbls.  The cause was an misordered BAT to perform the
calculations on, since the casted int values were added to the BAT, and
the example was carefully chosen to make sure the inverses for both
colums were true, e.g. after inserting, the BATs both had opposite
order.  We now leftjoin back with the intermediate selection set to
achieve identical order for both columns.  The exact/correct order is
lateron restored based on the input set.