alter_table_describe: remove redundant JDBC-based version of test
This test dates back to the time when mclient couldn't "describe"
tables, like JdbcClient.  Nowadays, mclient simply outperforms
JdbcClient on this, and JdbcClient shouldn't really be used any more.
So, remove the JDBC-based test in favour of the mclient-based one.
Incidently this fixes a problem on slow platforms, since JdbcClient
needs a lot of time to calculate the dependency graph of all
dependencies in order to find a suitable dumping order.