url_script_test: "fix" and provide stable outputs
Now getContent actually works, try to use it on a local file, such that
we don't need an internet connection to check if it works.

Approved all output of the operations, while most of them are wrong.
This is inherent to the ad-hoc parsing style employed.  Still approving,
since the output seems to be the intended output (although it doesn't
make much sense).