Client struct: register SQL and JAQL contexts directly
Instead of hiding the SQL and JAQL context pointers in some state
fields, create a field in the Client struct to store the pointers.  At
the moment SQL and JAQL both have a separate field, such that both
contexts can be loaded (and in use).

Adapted the SQL and JAQL code to set and retrieve the context pointers
from the sqlcontext and jaqlcontext fields from the Client struct.

Because SQL uses a backend struct and mvc struct, where the former has a
pointer to the latter, the SQL code was changed to only store the
backend struct pointer, and retrieve the mvc struct pointer via the
backend pointer.