MT_getrss: fix Darwin implementation
We have to use more non-documented voodoo.  It appears
TASK_BASIC_INFO doesn't return a useable/valid number for the
resident_size field.  TASK_BASIC_INFO_64 seems to, as seen with top.
Use mach_task_self() to avoid looking up the mach task by the self pid,
it may be faster like this, at least less error checking needs to be
done.
Further, the resident_size is returned in bytes, so we need not to
multiply by 1024.  The "sort-of" documentation mentions it should be the
number of pages (hence to be multiplied by the pagesize), but in reality
it is the number of bytes.