ResultsetResponse.getLine: fix possible divide by zero error
cacheSize can turn 0 when the amount of free memory becomes 0, with a
divide by zero as a result.
The free memory thing was completely bogus, as it was counting how much
memory was free, but used it to check how many rows of data we could
hold.  A row obviously requires more than just 2 bytes (UTF-16), so this
is flawed.  Instead, it's better just to rely on the JVM to do the right
thing with memory management, and hence not to keep on increasing the
cacheSize until memory seems to be short.  Instead we now just use a
fixed increment of the cacheSize, to fetch bigger blocks, this should
gain the efficiency aimed at, while not blowing up the memory
requirements to a rediculous amount either.

This fixes bug #3119.