Experimental OID list compression The intermediates of queries comprise of large OID sets. The views are a good method to compress them when a range is involved.
This batch of changes introduce an optimizer to inject a compression/
decompression scheme at runtime for oid,oid BATs. It is rough and
require some debugging on larger sets. Also optimization to ignore
badly compressable oid lists and avoidance of multi-decompression
could be studied.

The compression scheme itself is based on exploitation of local
properties, such as oid-ranges and small oid-sets.
Extension towards other datatypes and compression schemes (e.g.rle)
could be considered.

In general, (de)compression trades cpu from memory footprint,
which means that this scheme is most likely to be of use in
out-of-memory situations.