monetdbd: wait for servers to be fully started
This is a solution for bug #3134.

Sabaoth now makes a difference between starting and started.  mserver5
uses this, such that processing the write ahead log should delay the
state of the server becoming started.  This way, monetdbd can wait for a
server to become fully available.
Unfortunately our global lock in combination with a possible indefinite
wait is a potential killer of the entire system.