sabaoth: resolve race-like condition in determining state
Because GDKinit can take some time, GDKlockHome may have been called
while msab_registerStarting was not.  In this gap, any msab_getStatus
call would conclude the database to be running, since the GDK lockfile
is locked.  Since we cannot set anything before we actually know that we
can start the database to be started (e.g. not locked by another
process), we now flag when a database is done starting, instead of
flagging it is still busy starting up.
This resolves the problem reported in bug #3134, comment #7, that was
introduced in Jul2012-SP1.