configure: don't set PYTHONX to an absolute path for AC_PATH_PROG
AC_PATH_PROG will not perform its check if the variable is already set
to an absolute path.  Hence, don't set PYTHON2 and PYTHON3 to an
absolute path prior to checking for them.  I don't understand why it is
necessary to set PYTHONX in "python", since we already do that in the
"python2" and "python3" cases.  So this should just work if there is no
pythonX binary, but python is version X.