Re-establish fixed color map heuristics
A fixed color scheme is designed to keep the legend
consistent over multiple tomograms. Furthermore, groups
of instructions are kept within the same color group.
For less frequent or important operations, we re-use the color.

The map design is initially based on the relative frequency
of MAL instructions encountered in TPCH and those that
take significant time.

The built-in fixed color map can be over-ruled using a user
defined color map file, which consists of mod\tfcn\tcol lines using
the color dictionary as a frame of reference.

Use the adaptive scheme if you want to leave colorpicking to
the code itself.