SQLstatementIntern(): fixed bug that caused segfaults (double free) with SciQL use case:
We must not free (_DELETE()) the global variable stack (c->glb)
in SQLstatementIntern() (even if MSinitClientPrg() clears it unconditionally)
because some calling function might hold and restore a backup (pointer)
eventually resulting in a double-free.

Instead, also in SQLstatementIntern(), we keep a local backup of c->glb
before calling MSinitClientPrg() and restore it afterwards.