increase GDKLIBRARY: 128-bit database is incompatible with 64-bit server and v.v
Databases that do (might) contain 128-bit integers (hge/HUGEINT)
cannot be opened with a server that does not support 128-bit integers;
hence, we now record in the BBP.dir header, whether the database was
created with largest (supported) integer being 64-bit or 128-bit,
increase GDKLIBRARY, and check compatibility when opening BBP.dir.

For now, this is strict conservative check, mutually excluding
64-bit and 128-bit databases.

In fact, a server with 128-bit integer support could easily and
safely open and read a database that was created without 128-bit
integer support, upgrading it on-the-fly. However, this would also
require a proper upgrade of the SQL catalog (if applicable).
For now, we have neither the check nor the upgrade implemented;
hence, we refrain from opening such "incompatible" databases.

Likewise, a server without 128-bit integer support could open
and read a database that was created with 128-bit integer support,
provided it does neither contain 128-bit integer data (BATs),
nor a SQL catalog (with, e.g., function signatures that require
type HUGEINT). Without having the respective checks implemented,
and considering it a rather rare case, we also disallow this
combination.