controlrunner: fixed compilation with --enable-strict: check result of chdir()
Fabian says "Ignore the output. The idea is that the stuff below
will also fail, and therefore emit some error, but if that happens,
the world already is in such a bad shape that that most likely isn't
your biggest problem.
Hence a (void) probably does.
If not, a fake if. (exit(0) should be fine)."
(https://www.monetdb.org/pipermail/developers-list/2014-February/004238.html)