removed outdated hardware performance / events counter module / interface
While in general fine, this code uses old, possibly outdated libraries and interfaces
to provide access to the CPU hardware performance / events counters.
Given that, and since the code has not been used / tested for a long time,
we no remove it completely.

It remains in the attic / histrory for reference in case it should be
revived / reincarnated (and modernized!) in the future.